requirements.txt
discord.py
requests
aiohttp
python-dotenv
